fix(SeriesTest): size_t types different between win/linux

+++ SrcLib/io/fwNetworkIO/test/tu/src/SeriesTest.cpp
@@ -101,7 +101,8 @@ void SeriesTest::testSeries()
     CPPUNIT_ASSERT_EQUAL(dicomSeries->getTime(), seriesTime.toStdString());
     CPPUNIT_ASSERT_EQUAL(dicomSeries->getDescription(), seriesDescription.toStdString());
     CPPUNIT_ASSERT_EQUAL(dicomSeries->getModality(), modality.toStdString());
-    CPPUNIT_ASSERT_EQUAL(dicomSeries->getNumberOfInstances(), numberOfSeriesRelatedInstances.toULong());
+    CPPUNIT_ASSERT_EQUAL(dicomSeries->getNumberOfInstances(),
+                         static_cast<size_t>(numberOfSeriesRelatedInstances.toULong()));
 
     ::fwMedData::Patient::sptr patient = series->getPatient();
     CPPUNIT_ASSERT(patient);
